Allow recursively changing Path ownership
