+++ Sources/TrailBlazer/Open/OpenDirectory.swift
@@ -1,3 +1,9 @@
+#if os(Linux)
+import Glibc
+#else
+import Darwin
+#endif
+
 public typealias OpenDirectory = Open<DirectoryPath>
 
 extension Open: Sequence, IteratorProtocol where PathType: DirectoryPath {
@@ -15,5 +21,27 @@ extension Open: Sequence, IteratorProtocol where PathType: DirectoryPath {
 
     public func next() -> GenericPath? {
         return path.next()
+    }
+}
+
+public extension Open where PathType: DirectoryPath {
+    public func changeRecursive(owner uid: uid_t = ~0, group gid: gid_t = ~0) throws {
+        try change(owner: uid, group: gid)
+
+        for path in self {
+            if let dir = DirectoryPath(path) {
+                guard !["..", "."].contains(dir.lastComponent) else { continue }
+                try dir.changeRecursive(owner: uid, group: gid)
+            } else {
+                try path.change(owner: uid, group: gid)
+            }
+        }
+    }
+
+    public func changeRecursive(owner username: String? = nil, group groupname: String? = nil) throws {
+        let uid: uid_t = username != nil ? DirectoryPath.getUserInfo(username!)?.pw_uid ?? ~0 : ~0
+        let gid: gid_t = groupname != nil ? DirectoryPath.getGroupInfo(groupname!)?.gr_gid ?? ~0 : ~0
+
+        try changeRecursive(owner: uid, group: gid)
     }
 }
